move cube-rl inside pipelinerl package
